refactor: move game session fetch to gameApi
