+++ codenames-frontend/src/components/RoomLobby/RoomLobby.tsx
@@ -15,7 +15,11 @@ import { apiUrl, frontendUrl, socketUrl } from "../../config/api.tsx";
 import { getCookie, getUserId } from "../../shared/utils.tsx";
 import { io } from "socket.io-client";
 import { useToast } from "../Toast/ToastContext.tsx";
-
+import { getGameSession } from "../../api/gameApi.tsx";
+import {
+  GameSessionRoomLobbyDTO,
+  UserRoomLobbyDTO,
+} from "../../api/gameApi.tsx";
 /**
  * @returns {string} - The URL of the API.
  */
@@ -32,62 +36,6 @@ interface RoomLobbyProps {
 }
 
 /**
- * Enum for user status.
- * @enum {string}
- * @property {string} INACTIVE - The user is inactive.
- * @property {string} ACTIVE - The user is active.
- */
-enum UserStatus {
-  INACTIVE = "INACTIVE",
-  ACTIVE = "ACTIVE",
-}
-
-/**
- * Represents a user in the game session.
- * @typedef {Object} UserRoomLobbyDTO
- * @property {string} id - The unique identifier of the user.
- * @property {string} username - The username of the player.
- * @property {number} profilePic - The profile picture ID of the player.
- * @property {UserStatus} status - The status of the player (active/inactive).
- */
-interface UserRoomLobbyDTO {
-  id: string;
-  username: string;
-  profilePic: number;
-  status: UserStatus;
-}
-
-/**
- * Enum for session status.
- * @enum {string}
- * @property {string} CREATED - The session is created.
- * @property {string} LEADER_SELECTION - The session is in leader selection phase.
- * @property {string} IN_PROGRESS - The session is in progress.
- * @property {string} FINISHED - The session is finished.
- */
-enum SessionStatus {
-  CREATED = "CREATED",
-  LEADER_SELECTION = "LEADER_SELECTION",
-  IN_PROGRESS = "IN_PROGRESS",
-  FINISHED = "FINISHED",
-}
-
-/**
- * Represents a game session.
- * @typedef {Object} GameSessionRoomLobbyDTO
- * @property {SessionStatus} status - The current status of the session.
- * @property {string} gameName - The name of the game.
- * @property {number} maxPlayers - The maximum number of players allowed.
- * @property {UserRoomLobbyDTO[][]} connectedUsers - List of users in each team.
- */
-interface GameSessionRoomLobbyDTO {
-  status: SessionStatus;
-  gameName: string;
-  maxPlayers: number;
-  connectedUsers: UserRoomLobbyDTO[][];
-}
-
-/**
  * RoomLobby component.
  *
  * @param {RoomLobbyProps} props - The properties for the RoomLobby component.
@@ -101,7 +49,7 @@ const RoomLobby: React.FC<RoomLobbyProps> = ({ soundFXVolume }) => {
     useState<GameSessionRoomLobbyDTO | null>(null);
   const [redTeamPlayers, setRedTeamPlayers] = useState<UserRoomLobbyDTO[]>([]);
   const [blueTeamPlayers, setBlueTeamPlayers] = useState<UserRoomLobbyDTO[]>(
-    []
+    [],
   );
   const [isJoined, setIsJoined] = useState(false);
   const [isJoinedRed, setIsJoinedRed] = useState(false);
@@ -112,10 +60,9 @@ const RoomLobby: React.FC<RoomLobbyProps> = ({ soundFXVolume }) => {
   const [lobbyLink, setLobbyLink] = useState<string>("");
   const [isLinkIsleExpanded, setIsLinkIsleExpanded] = useState(false);
   const exampleLink = `${frontendUrl}/invite/${sessionStorage.getItem(
-    "gameId"
+    "gameId",
   )}`;
 
-  //TODO: move this fetch
   /**
    * Initializes the WebSocket connection and fetches the game session data.
    */
@@ -123,9 +70,8 @@ const RoomLobby: React.FC<RoomLobbyProps> = ({ soundFXVolume }) => {
     const storedGameId = sessionStorage.getItem("gameId");
 
     if (storedGameId) {
-      fetch(`${apiUrl}/api/game-session/${storedGameId}`)
-        .then((response) => response.json())
-        .then((data: GameSessionRoomLobbyDTO) => {
+      getGameSession(storedGameId)
+        .then((data) => {
           setGameSession({
             ...data,
           });
@@ -155,7 +101,7 @@ const RoomLobby: React.FC<RoomLobbyProps> = ({ soundFXVolume }) => {
     gameSocket.on("gameSessionUpdate", (updatedGameSessionJson: string) => {
       try {
         const updatedGameSession: GameSessionRoomLobbyDTO = JSON.parse(
-          updatedGameSessionJson
+          updatedGameSessionJson,
         );
 
         if (updatedGameSession.connectedUsers) {
@@ -204,7 +150,7 @@ const RoomLobby: React.FC<RoomLobbyProps> = ({ soundFXVolume }) => {
       {
         method: "POST",
         credentials: "include",
-      }
+      },
     );
 
     if (response.ok) {
@@ -239,7 +185,7 @@ const RoomLobby: React.FC<RoomLobbyProps> = ({ soundFXVolume }) => {
       {
         method: "POST",
         credentials: "include",
-      }
+      },
     );
 
     if (response.ok) {
@@ -274,7 +220,7 @@ const RoomLobby: React.FC<RoomLobbyProps> = ({ soundFXVolume }) => {
       //TODO: move this fetch
       const response = await fetch(
         `${apiUrl}/api/game-session/${storedGameId}/disconnect?userId=${userId}`,
-        { method: "DELETE", credentials: "include" }
+        { method: "DELETE", credentials: "include" },
       );
 
       if (!response.ok) {
@@ -305,7 +251,6 @@ const RoomLobby: React.FC<RoomLobbyProps> = ({ soundFXVolume }) => {
    * Starts the game session.
    */
   const startGame = async () => {
-
     const storedGameId = sessionStorage.getItem("gameId");
     if (!storedGameId) return;
 
@@ -317,7 +262,7 @@ const RoomLobby: React.FC<RoomLobbyProps> = ({ soundFXVolume }) => {
       //TODO: move this fetch
       const response = await fetch(
         `${apiUrl}/api/game-session/${storedGameId}/start`,
-        { method: "POST" }
+        { method: "POST" },
       );
       if (response.ok) {
         navigate("/choose-leader");
@@ -359,7 +304,7 @@ const RoomLobby: React.FC<RoomLobbyProps> = ({ soundFXVolume }) => {
 
       setTimeout(() => {
         setNotifications((prevNotifications) =>
-          prevNotifications.filter((notification) => notification.id !== id)
+          prevNotifications.filter((notification) => notification.id !== id),
         );
       }, 500);
     }
@@ -378,7 +323,7 @@ const RoomLobby: React.FC<RoomLobbyProps> = ({ soundFXVolume }) => {
           soundFXVolume={soundFXVolume}
           onClick={removePlayer}
         >
-          <img src={backButton} alt="Back" className="btn-arrow-back"/>
+          <img src={backButton} alt="Back" className="btn-arrow-back" />
         </Button>
         <span className="room-form-label">{t("game-lobby")}</span>
         <div className="room-lobby-divider">
@@ -415,7 +360,7 @@ const RoomLobby: React.FC<RoomLobbyProps> = ({ soundFXVolume }) => {
                 }`}
                 onClick={handleLobbyLinkIsleUnroll}
               >
-                <img src={messageIcon} alt="Link" className="isle-image"/>
+                <img src={messageIcon} alt="Link" className="isle-image" />
                 <p className="isle-title">{t("invite-friends")}</p>
                 <p className="isle-text">{t("invite-friends-text")}</p>
                 <p className="isle-fields">
